common-treble: Switch to CAF boot HAL 1.1
AOSP switched up the default implementation of the bootctrl 1.1 HAL from
GPT flags to the `/misc` partition (`-impl` lib doesn't load the
bootctrl libhardware module anymore but delegates to
`libboot_control.so`) [1], which is incorrect on our Qualcomm boards.
As we'd still like to switch to the 1.1 interface for FCM5
compatibility, and would like to drop the legacy libhardware module (see
`bootctl.<platform>` in the respective platform directories, and its
implementation in `hardware/qcom/bootctrl`) we'll use the updated CAF
repository directly.

Note that we still include AOSP's 1.1 passthrough library loader for
lack of a startable service in the CAF repo.  As usual, this provides us
with the right init .rc and vintf fragment, and allows starting both a
1.0 and 1.1 passthrough (`-impl`) library.  It is this `-impl` library
that is custom to QTI and calls into their custom
`libboot_control_qti.so`.

[1]: https://github.com/sonyxperiadev/bug_tracker/issues/740#issuecomment-1016325638

Fixes: 1a91f1b6 ("Upgrade IBootControl HAL to 1.1")
Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>