common : drop thermal 1.0 which is deprecated in FCM 4
Android 9 and lower use a polling interface defined in Thermal HAL 1.0 to get temperature readings. This HAL allowed the Android framework and other trusted clients, such as a device manufacturer's HAL, to read the current temperature and product-policy specific throttling and shutdown thresholds for each sensor through the same API.

Android 10 introduced a thermal system in the Android framework and a new version of the HAL, Thermal HAL 2.0, that abstracts the interface to the thermal subsystem hardware devices. The hardware interface includes temperature sensors and thermistors for the skin, battery, GPU, CPU, and USB port..