hardware: health: Add fallback to `battery` if `bms` is not provided
Some devices do not have a BMS interface, so information about
the battery status is obtained directly from the battery interface.
Move the definition of charge_full and cycle_count and variables
to the corresponding constructors and add a fallback mechanism,
in case of no BMS, use the battery sysfs power_supply path.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>