[PATCH] common: Include Wi-FI HAL soong namespace
WiFi HAL migrated to Android.bp and in order for all its components
to be built we need to specify the namespace. AOSP provides two
versions of HAL. A very old legacy version and a fresh one, which they
protected with the BOARD_WLAN_CHIP[2] flag which corresponds to the
value wcn6740.

[1] https://android.googlesource.com/platform/hardware/qcom/wlan/+/eb93a5121d88779e93a2b361ac0ddac1e98c02e8
[2] https://android.googlesource.com/platform/hardware/qcom/wlan/+/e6de8ddaa1c92487f4073952ff08c124e3ba87ea

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>